refactor(rendering): estrai envelope_extractor da ScoreVisualizer (issue #150)
Sposta _get_stream_envelopes / _base_param_name / _get_voice_offset_envelopes
in rendering/envelope_extractor.py come funzioni di modulo matplotlib-free
(get_stream_envelopes / base_param_name / get_voice_offset_envelopes). I metodi
del visualizer diventano wrapper sottili che leggono self.config e delegano:
nessuna regressione (suite test_score_visualizer* verde). Prepara la single
source of truth condivisa col futuro SVExporter.
